Change result type in main method

+++ src/main/java/pl/sda/calculator/Main.java
@@ -4,7 +4,7 @@ public class Main {
 
     public static void main(String... args) {
         Calculator calculator = new Calculator();
-        double result = calculator.add(2, 3);
+        int result = calculator.add(2, 3);
         System.out.println(result);
     }
 }
